pyproject.toml
[project]
name = "langgraph-memory-redis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-core>=1.34.0",
    "azure-identity>=1.23.0",
    "dill>=0.4.0",
    "lagom>=2.7.5",
    "langchain-openai>=0.3.23",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
    "redis>=6.2.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.2.0",
    "pyright>=1.1.402",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.2.1",
    "pytest-mock>=3.14.1",
    "ruff>=0.11.13",
]

[tool.ruff]
exclude = ["typings"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "TID"]

[tool.ruff.format]
docstring-code-format = true

[tool.pyright]
venvPath = "."
venv = ".venv"
exclude = ["typings", ".venv", ".vscode", ".git"]

[tool.coverage.run]
omit = [
    "*/__init__.py",
    "langgraph_memory_redis/hosting.py",
    "langgraph_memory_redis/protocols/*",
    "langgraph_memory_redis/main.py",
    "langgraph_memory_redis/server.py"
]
